debug

+++ utils.py
@@ -64,7 +64,7 @@ def train_loop(dataloader, model, loss_fn, optimizer, device):
         loss.backward()
         optimizer.step()
         total_loss += loss.item()
-        correct_cases += torch.sum(torch.squeeze(pred) == y).item()/X.shape[0]
+        correct_cases += torch.sum(torch.squeeze(pred) == y.float()).item()/X.shape[0]
         # if cnt % 100:
         #    print(f"Current avg loss:{total_loss/(cnt+1)}\n")
         # cnt += 1   
@@ -86,7 +86,7 @@ def test_loop(dataloader, model, loss_fn, device):
             pred = model(X)
             # squeeze to match the dimensions of pred and y
             total_loss += loss_fn(torch.squeeze(pred), y.float()).item()
-            correct_cases += torch.sum(torch.squeeze(pred) == y).item()/X.shape[0]
+            correct_cases += torch.sum(torch.squeeze(pred) == y.float()).item()/X.shape[0]
 
     return total_loss/num_batches, correct_cases/num_batches
